Start writing the JavaScript standard library
